Fixes RNG
