+++ Projects/Server/Utilities/Random.cs
@@ -1,18 +1,24 @@
-/***************************************************************************
- *
- *   This program is free software; you can redistribute it and/or modify
- *   it under the terms of the GNU General Public License as published by
- *   the Free Software Foundation; either version 2 of the License, or
- *   (at your option) any later version.
- *
- ***************************************************************************/
+/*************************************************************************
+ * ModernUO                                                              *
+ * Copyright (C) 2019 - ModernUO Development Team                        *
+ * Email: hi@modernuo.com                                                *
+ * File: Random.cs - Created: 2019/12/30 - Updated: 2019/12/30           *
+ *                                                                       *
+ * This program is free software: you can redistribute it and/or modify  *
+ * it under the terms of the GNU General Public License as published by  *
+ * the Free Software Foundation, either version 3 of the License, or     *
+ * (at your option) any later version.                                   *
+ *                                                                       *
+ * This program is distributed in the hope that it will be useful,       *
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ * GNU General Public License for more details.                          *
+ *                                                                       *
+ * You should have received a copy of the GNU General Public License     *
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
+ *************************************************************************/
 
 using System;
-using System.IO;
-using System.Runtime.InteropServices;
-using System.Security.Cryptography;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace Server
 {
@@ -21,21 +27,7 @@ namespace Server
   /// </summary>
   public static class RandomImpl
   {
-    private static readonly IRandomImpl _Random;
-    // private static readonly IRandomImpl _SecureRandom;
-
-    static RandomImpl()
-    {
-      if (Core.Unix && File.Exists("rdrand.so") || !Core.Unix && File.Exists("rdrand.dll"))
-        _Random = new RDRand64();
-
-      if (_Random == null || _Random is IHardwareRNG rng && !rng.IsSupported())
-        _Random = new CSPRandom();
-    }
-
-    public static bool IsHardwareRNG => _Random is IHardwareRNG;
-
-    public static Type Type => _Random.GetType();
+    private static readonly IRandom _Random = new Xoshiro256PlusPlus();
 
     public static uint Next(uint c) => _Random.Next(c);
 
@@ -46,7 +38,38 @@ namespace Server
     public static double NextDouble() => _Random.NextDouble();
   }
 
-  public interface IRandomImpl
+  public static class SecureRandomImpl
+  {
+    private static readonly IRandom _Random;
+
+    static SecureRandomImpl()
+    {
+      try
+      {
+        _Random = new DRng64();
+        if (_Random == null || _Random is IHardwareRNG rng && !rng.IsSupported())
+          _Random = new CSPRng();
+      }
+      catch (Exception)
+      {
+        _Random = new CSPRng();
+      }
+    }
+
+    public static bool IsHardwareRNG => _Random is IHardwareRNG;
+
+    public static string Name => _Random.GetType().Name;
+
+    public static uint Next(uint c) => _Random.Next(c);
+
+    public static bool NextBool() => _Random.NextBool();
+
+    public static void NextBytes(Span<byte> b) => _Random.NextBytes(b);
+
+    public static double NextDouble() => _Random.NextDouble();
+  }
+
+  public interface IRandom
   {
     uint NextUInt32();
     uint Next(uint n);
@@ -62,7 +85,7 @@ namespace Server
     bool IsSupported();
   }
 
-  public abstract class BaseRandom : IRandomImpl
+  public abstract class BaseRandom : IRandom
   {
     internal abstract void GetBytes(Span<byte> b);
     internal abstract void GetBytes(byte[] b, int offset, int count);
@@ -109,157 +132,5 @@ namespace Server
 
       return (double)r / 9007199254740992;
     }
-  }
-
-  public sealed class CSPRandom : BaseRandom
-  {
-    private static int BUFFER_SIZE = 0x4000;
-    private static int LARGE_REQUEST = 0x40;
-    private byte[] _Buffer = new byte[BUFFER_SIZE];
-    private RNGCryptoServiceProvider _CSP = new RNGCryptoServiceProvider();
-
-    private ManualResetEvent _filled = new ManualResetEvent(false);
-
-    private int _Index;
-
-    private object _sync = new object();
-
-    private byte[] _Working = new byte[BUFFER_SIZE];
-
-    public CSPRandom()
-    {
-      _CSP.GetBytes(_Working);
-      Task.Run(Fill);
-    }
-
-    public override void NextBytes(Span<byte> b)
-    {
-      int c = b.Length;
-
-      if (c >= LARGE_REQUEST)
-      {
-        lock (_sync)
-        {
-          _CSP.GetBytes(b);
-        }
-
-        return;
-      }
-
-      GetBytes(b);
-    }
-
-    private void CheckSwap(int c)
-    {
-      if (_Index + c < BUFFER_SIZE)
-        return;
-
-      _filled.WaitOne();
-
-      byte[] b = _Working;
-      _Working = _Buffer;
-      _Buffer = b;
-      _Index = 0;
-
-      _filled.Reset();
-
-      Task.Run(Fill);
-    }
-
-    private void Fill()
-    {
-      _CSP.GetBytes(_Buffer);
-      _filled.Set();
-    }
-
-    internal override void GetBytes(Span<byte> b)
-    {
-      int c = b.Length;
-
-      lock (_sync)
-      {
-        CheckSwap(c);
-        _Working.AsSpan(0, c).CopyTo(b);
-        _Index += c;
-      }
-    }
-
-    internal override void GetBytes(byte[] b, int offset, int count)
-    {
-      GetBytes(b.AsSpan(offset, count));
-    }
-
-    public override byte NextByte()
-    {
-      lock (_sync)
-      {
-        CheckSwap(1);
-        return _Working[_Index++];
-      }
-    }
-  }
-
-  // public sealed class RDRandUnix : BaseRandom, IHardwareRNG
-  // {
-  //   [DllImport("rdrand.so")]
-  //   internal static extern RDRandError rdrand_32(ref uint rand, bool retry);
-  //
-  //   [DllImport("rdrand.so")]
-  //   internal static extern unsafe RDRandError rdrand_get_bytes(int n, byte* buffer);
-  //
-  //   public bool IsSupported()
-  //   {
-  //     uint r = 0;
-  //     return rdrand_32(ref r, true) == RDRandError.Success;
-  //   }
-  //
-  //   internal override unsafe void GetBytes(Span<byte> b)
-  //   {
-  //     fixed(byte* ptr = b)
-  //       rdrand_get_bytes(b.Length, ptr);
-  //   }
-  //
-  //   internal override void GetBytes(byte[] b, int offset, int count)
-  //   {
-  //     GetBytes(b.AsSpan().Slice(offset, count));
-  //   }
-  // }
-
-  public sealed class RDRand64 : BaseRandom, IHardwareRNG
-  {
-    [DllImport("rdrand")]
-    internal static extern RDRandError rdrand_64(ref ulong rand, bool retry);
-
-    [DllImport("rdrand")]
-    internal static extern unsafe RDRandError rdrand_get_bytes(int n, byte* buffer);
-
-    public bool IsSupported()
-    {
-      ulong r = 0;
-      return rdrand_64(ref r, true) == RDRandError.Success;
-    }
-
-    internal override unsafe void GetBytes(Span<byte> b)
-    {
-      fixed (byte* ptr = b)
-        rdrand_get_bytes(b.Length, ptr);
-    }
-
-    internal override void GetBytes(byte[] b, int offset, int count)
-    {
-      GetBytes(b.AsSpan(offset, count));
-    }
-  }
-
-  public enum RDRandError
-  {
-    Unknown = -4,
-    Unsupported = -3,
-    Supported = -2,
-    NotReady = -1,
-
-    Failure = 0,
-
-    Success = 1
   }
 }
