+++ Projects/Server/Utilities/Xoshiro256PlusPlus.cs
@@ -1,9 +1,32 @@
+/*************************************************************************
+ * ModernUO                                                              *
+ * Copyright (C) 2019 - ModernUO Development Team                        *
+ * Email: hi@modernuo.com                                                *
+ * File: Xoshiro256PlusPlus.cs                                           *
+ * Created: 2019/12/29 - Updated: 2019/12/30                             *
+ *                                                                       *
+ * This program is free software: you can redistribute it and/or modify  *
+ * it under the terms of the GNU General Public License as published by  *
+ * the Free Software Foundation, either version 3 of the License, or     *
+ * (at your option) any later version.                                   *
+ *                                                                       *
+ * This program is distributed in the hope that it will be useful,       *
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ * GNU General Public License for more details.                          *
+ *                                                                       *
+ * You should have received a copy of the GNU General Public License     *
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
+ *************************************************************************/
+
+#nullable enable
+
 using System;
 using System.Runtime.CompilerServices;
 
 namespace Server
 {
-  public class Xoshiro256PlusPlus : IRandomImpl
+  public class Xoshiro256PlusPlus : IRandom
   {
     private ulong _s0, _s1, _s2, _s3;
 
@@ -33,9 +56,9 @@ namespace Server
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public uint Next(uint max)
     {
-      if ( max <= 1u << 12 )
+      if (max <= 1u << 12)
       {
-        if ( max == 0 ) throw new ArgumentOutOfRangeException();
+        if (max == 0) throw new ArgumentOutOfRangeException();
         return (uint)(((ulong)NextUInt32() * max) >> 32);
       }
 
@@ -44,7 +67,7 @@ namespace Server
       {
         r = NextUInt32();
         v = r % max;
-      } while(r - v > limit);
+      } while (r - v > limit);
 
       return v;
     }
@@ -53,7 +76,8 @@ namespace Server
     public uint NextUInt32() => (uint)NextUInt64();
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public ulong NextUInt64() {
+    public ulong NextUInt64()
+    {
       ulong r1 = (_s1 << 2) + _s1;
       ulong r2 = (r1 << 7) | (r1 >> 57);
       ulong rslt = (r2 << 3) + r2;
@@ -75,16 +99,16 @@ namespace Server
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public ulong Next(ulong max)
     {
-      if ( max <= uint.MaxValue ) return Next((uint)max);
+      if (max <= uint.MaxValue) return Next((uint)max);
 
-      if ( max <= 1ul << 38 ) return ((NextUInt32() * max >> 32) + (NextUInt32() & ((1u << 26) - 1)) * max) >> 26;
+      if (max <= 1ul << 38) return ((NextUInt32() * max >> 32) + (NextUInt32() & ((1u << 26) - 1)) * max) >> 26;
 
       ulong r, v, limit = (ulong)-(long)max;
       do
       {
         r = NextUInt64();
         v = r % max;
-      } while(r - v > limit);
+      } while (r - v > limit);
 
       return v;
     }
@@ -100,11 +124,11 @@ namespace Server
 
       int i = 0;
 
-      fixed(byte* pBuffer = b)
+      fixed (byte* pBuffer = b)
       {
         ulong* pULong = (ulong*)pBuffer;
 
-        for(int bound = b.Length / 8; i < bound; i++)
+        for (int bound = b.Length / 8; i < bound; i++)
         {
           ulong r1 = (s1 << 2) + s1;
           ulong r2 = (r1 << 7) | (r1 >> 57);
@@ -124,7 +148,7 @@ namespace Server
 
       i *= 8;
 
-      if(i < b.Length)
+      if (i < b.Length)
       {
         ulong r1 = (s1 << 2) + s1;
         ulong r2 = (r1 << 7) | (r1 >> 57);
@@ -157,10 +181,10 @@ namespace Server
     public double NextDouble() => (NextUInt64() >> 11) * (1.0 / (1ul << 53));
 
     private static readonly ulong[] JUMP =
-      { 0x180ec6d33cfd0aba, 0xd5a61266f0c9392c, 0xa9582618e03fc9aa, 0x39abdc4529b1661c };
+      {0x180ec6d33cfd0aba, 0xd5a61266f0c9392c, 0xa9582618e03fc9aa, 0x39abdc4529b1661c};
 
     private static readonly ulong[] LONG_JUMP =
-      { 0x76e15d3efefdcbbf, 0xc5004e441c522fb3, 0x77710069854ee241, 0x39109bb02acbe635 };
+      {0x76e15d3efefdcbbf, 0xc5004e441c522fb3, 0x77710069854ee241, 0x39109bb02acbe635};
 
     public void Jump() => Jump(JUMP);
 
@@ -173,8 +197,8 @@ namespace Server
       ulong s2 = 0;
       ulong s3 = 0;
 
-      for(int i = 0; i < jumps.Length; i++)
-      for(int b = 0; b < 64; b++)
+      for (int i = 0; i < jumps.Length; i++)
+      for (int b = 0; b < 64; b++)
       {
         if ((jumps[i] & 1ul << b) != 0)
         {
@@ -183,6 +207,7 @@ namespace Server
           s2 ^= _s2;
           s3 ^= _s3;
         }
+
         NextUInt64();
       }
 
@@ -204,6 +229,28 @@ namespace Server
       Xoshiro256PlusPlus rng = new Xoshiro256PlusPlus(_s0, _s1, _s2, _s3);
       rng.LongJump();
       return rng;
+    }
+  }
+
+  public class SplitMix64
+  {
+    private ulong x;
+
+    public SplitMix64(ulong seed) => x = seed;
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public ulong Next()
+    {
+      ulong z = x += 0x9e3779b97f4a7c15;
+      z = (z ^ (z >> 30)) * 0xbf58476d1ce4e5b9;
+      z = (z ^ (z >> 27)) * 0x94d049bb133111eb;
+      return z ^ (z >> 31);
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public void FillArray(ulong[] arr)
+    {
+      for (int i = 0; i < arr.Length; i++) arr[i] = Next();
     }
   }
 }
